% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cilab-package.R
\docType{package}
\name{cilab-package}
\alias{cilab}
\alias{cilab-package}
\title{cilab: CI tests on Gitub}
\description{
Dummy package to test CI on Gitub.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/EricMarcon/cilab}
}

}
\author{
\strong{Maintainer}: Eric Marcon \email{e.marcon@free.fr} (\href{https://orcid.org/0000-0002-5249-321X}{ORCID})

}
\keyword{internal}
